+++ tests/testthat/test-cir.R
@@ -8,12 +8,12 @@ test_that("cir_query()", {
   expect_equal(cir_query("3380-34-5", 'stdinchikey', resolver = 'cas_number', verbose = FALSE)[[1]],
             "InChIKey=XEFQLINVKFYRCS-UHFFFAOYSA-N")
   expect_true(length(cir_query('Triclosan', 'cas', verbose = FALSE)[[1]]) > 1)
-  expect_message(cir_query("acetic acid", "mw", first = TRUE))
-  expect_length(cir_query('Triclosan', 'cas', choices = 1, verbose = FALSE)[[1]], 1)
+  expect_message(cir_query("acetic acid", "mw", match = "first"))
+  expect_length(cir_query('Triclosan', 'cas', match = "first", verbose = FALSE)[[1]], 1)
   expect_length(cir_query(c('Triclosan', 'Aspirin'), 'cas', verbose = FALSE), 2)
 
-  skip("I have no clue why this one fails on R CMD check.  It works when run in the console!")
-  expect_equivalent(cir_query('acetic acid', 'mw', choices = 1), c(`acetic acid` = 60.0524))
+  # skip("I have no clue why this one fails on R CMD check.  It works when run in the console!")
+  expect_equivalent(cir_query('acetic acid', 'mw', match = "first"), c(`acetic acid` = 60.0524))
 
 })